refactor: move utility functions from make_negative_controls to settings.py under utils
