+++ src/readii_2_roqc/readii/make_negative_controls.py
@@ -14,92 +14,7 @@ from readii.negative_controls_refactor import NegativeControlManager
 from readii.process.config import get_full_data_name
 from readii.utils import logger
 from readii_2_roqc.utils.metadata import get_masked_image_metadata
-
-REGIONS = {'full', 'roi', 'non_roi'}
-PERMUTATIONS = {'shuffled', 'sampled', 'randomized'}
-CROP = {'cube', 'bbox', 'centroid'}
-
-def check_setting_superset(setting_list: set, 
-                           setting_request: set | list | None):
-    """
-    Check if the requested settings are all in the global settings list for READII.
-
-    Parameters
-    ----------
-    setting_list: set
-        Set of allowed values for a READII setting. Should be set using the global sets in this file.
-    setting_request: set |
-    """
-    if setting_request == [] or setting_request is None:
-        return True
-    if not setting_list.issuperset(setting_request):
-        message = f"Requested settings ({setting_request}) is not a subset of the allowed settings ({setting_list})."
-        logger.error(message)
-        raise ValueError(message)
-
-    else:
-        return True
-
-
-
-def get_readii_settings(dataset_config: dict) -> tuple[list, list, list]:
-    """Extract READII settings from a configuration dictionary.
-    
-    Parameters
-    ----------
-    dataset_config : dict
-        Configuration dictionary read in with `loadImageDatasetConfig` containing READII settings
-    
-    Returns
-    -------
-    tuple
-        A tuple containing:
-        - regions: list of regions to process
-        - permutations: list of permutations to apply
-        - crop: list of crop settings
-    """
-    readii_config = dataset_config['READII']
-    if 'IMAGE_TYPES' not in readii_config:
-        message = "READII configuration must contain 'IMAGE_TYPES'."
-        logger.error(message)
-        raise KeyError(message)
-    
-    regions = readii_config['IMAGE_TYPES']['regions']
-    # Confirm requested regions are available settings
-    assert check_setting_superset(REGIONS, regions)
-        
-    permutations = readii_config['IMAGE_TYPES']['permutations']
-    # Confirm requested permutations are available settings
-    assert check_setting_superset(PERMUTATIONS, permutations)
-
-    crop = readii_config['IMAGE_TYPES']['crop']
-    if crop is not None and crop != []:
-        # Confirm requested crop is an available setting
-        assert check_setting_superset(CROP, crop)
-        # Get single crop value out of list format
-        crop = crop[0]
-    else:
-        crop = ""
-
-    resize = readii_config['IMAGE_TYPES']['resize']
-    if resize is None:
-        resize = []
-    elif isinstance(resize, list):
-        match len(resize):
-            case 1: resize = resize[0]
-            case 3: resize = resize
-            case 0: resize
-            case _: 
-                message = f"READII resize must be a single int, or list of three ints (e.g. [50, 50, 50]). Current value: {resize}"
-                logger.error(message)
-                raise TypeError(message)
-    elif not isinstance(resize, int):
-        message = f"READII resize must be a single int, or list of three values (e.g. [50, 50, 50]). Current value: {resize}"
-        logger.error(message)
-        raise TypeError(message)
-
-    return regions, permutations, crop, resize
-
+from readii_2_roqc.utils.settings import get_readii_settings, check_setting_superset
 
 
 def save_out_negative_controls(nifti_writer: NIFTIWriter,
